[FOLD] more splitting
