+++ Builds/VisualStudio2013/RippleD.vcxproj.filters
@@ -412,11 +412,11 @@
     <Filter Include="ripple\test">
       <UniqueIdentifier>{B25F5854-84AE-1CBD-DFFC-6515DD055652}</UniqueIdentifier>
     </Filter>
-    <Filter Include="ripple\test\impl">
-      <UniqueIdentifier>{CF7F0C3F-3D61-7764-BA8B-5FF38018425C}</UniqueIdentifier>
-    </Filter>
     <Filter Include="ripple\test\jtx">
       <UniqueIdentifier>{A21A3B94-5C44-3746-4F10-6FF8FF990CE3}</UniqueIdentifier>
+    </Filter>
+    <Filter Include="ripple\test\jtx\impl">
+      <UniqueIdentifier>{27D70888-7145-691C-0E0A-E511EB3A80A2}</UniqueIdentifier>
     </Filter>
     <Filter Include="ripple\unity">
       <UniqueIdentifier>{5DB3CD0B-B361-B301-9562-697CA8A52B68}</UniqueIdentifier>
@@ -4086,40 +4086,61 @@
     <ClInclude Include="..\..\src\ripple\shamap\TreeNodeCache.h">
       <Filter>ripple\shamap</Filter>
     </ClInclude>
-    <ClInclude Include="..\..\src\ripple\test\Account.h">
-      <Filter>ripple\test</Filter>
+    <ClInclude Include="..\..\src\ripple\test\jtx\Account.h">
+      <Filter>ripple\test\jtx</Filter>
     </ClInclude>
-    <ClInclude Include="..\..\src\ripple\test\amounts.h">
-      <Filter>ripple\test</Filter>
-    </ClInclude>
-    <ClInclude Include="..\..\src\ripple\test\Env.h">
-      <Filter>ripple\test</Filter>
-    </ClInclude>
-    <ClCompile Include="..\..\src\ripple\test\impl\Account.cpp">
-      <Filter>ripple\test\impl</Filter>
-    </ClCompile>
-    <ClCompile Include="..\..\src\ripple\test\impl\amounts.cpp">
-      <Filter>ripple\test\impl</Filter>
-    </ClCompile>
-    <ClCompile Include="..\..\src\ripple\test\impl\Env.cpp">
-      <Filter>ripple\test\impl</Filter>
-    </ClCompile>
-    <ClCompile Include="..\..\src\ripple\test\impl\Env_test.cpp">
-      <Filter>ripple\test\impl</Filter>
-    </ClCompile>
-    <ClCompile Include="..\..\src\ripple\test\impl\JTx.cpp">
-      <Filter>ripple\test\impl</Filter>
-    </ClCompile>
-    <ClInclude Include="..\..\src\ripple\test\JTx.h">
-      <Filter>ripple\test</Filter>
+    <ClInclude Include="..\..\src\ripple\test\jtx\amounts.h">
+      <Filter>ripple\test\jtx</Filter>
     </ClInclude>
     <ClInclude Include="..\..\src\ripple\test\jtx\any.h">
       <Filter>ripple\test\jtx</Filter>
     </ClInclude>
+    <ClInclude Include="..\..\src\ripple\test\jtx\Env.h">
+      <Filter>ripple\test\jtx</Filter>
+    </ClInclude>
+    <ClCompile Include="..\..\src\ripple\test\jtx\impl\Account.cpp">
+      <Filter>ripple\test\jtx\impl</Filter>
+    </ClCompile>
+    <ClCompile Include="..\..\src\ripple\test\jtx\impl\amounts.cpp">
+      <Filter>ripple\test\jtx\impl</Filter>
+    </ClCompile>
+    <ClCompile Include="..\..\src\ripple\test\jtx\impl\Env.cpp">
+      <Filter>ripple\test\jtx\impl</Filter>
+    </ClCompile>
+    <ClCompile Include="..\..\src\ripple\test\jtx\impl\Env_test.cpp">
+      <Filter>ripple\test\jtx\impl</Filter>
+    </ClCompile>
+    <ClCompile Include="..\..\src\ripple\test\jtx\impl\json.cpp">
+      <Filter>ripple\test\jtx\impl</Filter>
+    </ClCompile>
+    <ClCompile Include="..\..\src\ripple\test\jtx\impl\JTx.cpp">
+      <Filter>ripple\test\jtx\impl</Filter>
+    </ClCompile>
+    <ClCompile Include="..\..\src\ripple\test\jtx\impl\multisign.cpp">
+      <Filter>ripple\test\jtx\impl</Filter>
+    </ClCompile>
+    <ClCompile Include="..\..\src\ripple\test\jtx\impl\owners.cpp">
+      <Filter>ripple\test\jtx\impl</Filter>
+    </ClCompile>
+    <ClCompile Include="..\..\src\ripple\test\jtx\impl\ticket.cpp">
+      <Filter>ripple\test\jtx\impl</Filter>
+    </ClCompile>
     <ClInclude Include="..\..\src\ripple\test\jtx\json.h">
       <Filter>ripple\test\jtx</Filter>
     </ClInclude>
+    <ClInclude Include="..\..\src\ripple\test\jtx\JTx.h">
+      <Filter>ripple\test\jtx</Filter>
+    </ClInclude>
+    <ClInclude Include="..\..\src\ripple\test\jtx\multisign.h">
+      <Filter>ripple\test\jtx</Filter>
+    </ClInclude>
+    <ClInclude Include="..\..\src\ripple\test\jtx\owners.h">
+      <Filter>ripple\test\jtx</Filter>
+    </ClInclude>
     <ClInclude Include="..\..\src\ripple\test\jtx\tags.h">
+      <Filter>ripple\test\jtx</Filter>
+    </ClInclude>
+    <ClInclude Include="..\..\src\ripple\test\jtx\ticket.h">
       <Filter>ripple\test\jtx</Filter>
     </ClInclude>
     <ClCompile Include="..\..\src\ripple\unity\app_ledger.cpp">
