+++ src/ripple/test/jtx/impl/Env.cpp
@@ -18,7 +18,7 @@
 //==============================================================================
 
 #include <BeastConfig.h>
-#include <ripple/test/Env.h>
+#include <ripple/test/jtx/Env.h>
 #include <ripple/app/paths/FindPaths.h>
 #include <ripple/app/tx/TransactionEngine.h>
 #include <ripple/basics/Slice.h>
